+++ bitcoin-jsonrpc-client/bitcoin-jsonrpc-client-example-application/src/main/java/org/tbk/bitcoin/jsonrpc/example/BitcoinJsonrpcClientExampleApplicationConfig.java
@@ -2,12 +2,12 @@ package org.tbk.bitcoin.jsonrpc.example;
 
 import com.google.common.base.Stopwatch;
 import lombok.extern.slf4j.Slf4j;
+import org.bitcoinj.core.Address;
 import org.bitcoinj.core.Block;
-import org.bitcoinj.core.Coin;
 import org.bitcoinj.core.Sha256Hash;
+import org.bitcoinj.params.RegTestParams;
 import org.consensusj.bitcoin.json.pojo.BlockChainInfo;
 import org.consensusj.bitcoin.json.pojo.NetworkInfo;
-import org.consensusj.bitcoin.json.pojo.UnspentOutput;
 import org.consensusj.bitcoin.jsonrpc.BitcoinClient;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.context.annotation.Bean;
@@ -15,6 +15,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 import org.springframework.scheduling.TaskScheduler;
 import org.springframework.scheduling.annotation.EnableScheduling;
+import org.tbk.bitcoin.jsonrpc.cache.CacheFacade;
 
 import java.io.IOException;
 import java.time.Duration;
@@ -27,7 +28,9 @@ class BitcoinJsonrpcClientExampleApplicationConfig {
 
     @Bean
     @Profile("!test")
-    CommandLineRunner printBitcoinNodeInfo(TaskScheduler scheduler, BitcoinClient bitcoinClient) {
+    CommandLineRunner printBitcoinNodeInfo(TaskScheduler scheduler,
+                                           BitcoinClient bitcoinClient,
+                                           CacheFacade cacheFacade) {
         Runnable task = () -> {
             Stopwatch stopwatch = Stopwatch.createStarted();
             try {
@@ -48,22 +51,14 @@ class BitcoinJsonrpcClientExampleApplicationConfig {
                 log.info("'getNetworkInfo' after {}", stopwatch);
                 log.info("Version: {} ({})", networkInfo.getVersion(), networkInfo.getSubVersion());
 
-                Coin balance = bitcoinClient.getBalance();
-                log.info("'getBalance' after {}", stopwatch);
-                log.info("Balance: {}", balance);
-
                 Sha256Hash genesisBlockHash = bitcoinClient.getBlockHash(0);
-                Block block0 = bitcoinClient.getBlock(genesisBlockHash);
+                Block block0 = cacheFacade.block().getUnchecked(genesisBlockHash);
                 log.info("'getBlock(0)' after {}", stopwatch);
                 log.info("block0 Time: {}", block0.getTime());
 
-                Block blockBest = bitcoinClient.getBlock(blockChainInfo.getBestBlockHash());
+                Block blockBest = cacheFacade.block().getUnchecked(blockChainInfo.getBestBlockHash());
                 log.info("'getBlock(best)' after {}", stopwatch);
                 log.info("blockBest Time: {}", blockBest.getTime());
-
-                final List<UnspentOutput> unspentOutputs = bitcoinClient.listUnspent();
-                log.info("'listUnspent' after {}", stopwatch);
-                log.info("UnspentOutputs: {}", unspentOutputs);
             } catch (IOException e) {
                 log.error("", e);
             } finally {
@@ -72,5 +67,39 @@ class BitcoinJsonrpcClientExampleApplicationConfig {
         };
 
         return args -> scheduler.scheduleWithFixedDelay(task, Duration.ofSeconds(10L));
+    }
+
+    @Bean
+    @Profile("!test")
+    CommandLineRunner printCacheStatsInfo(TaskScheduler scheduler, CacheFacade cacheFacade) {
+        Runnable task = () -> {
+            log.info("Block cache: {}", cacheFacade.block().stats());
+            log.info("Block Info cache: {}", cacheFacade.blockInfo().stats());
+            log.info("Transaction cache: {}", cacheFacade.tx().stats());
+            log.info("Raw Transaction Info cache: {}", cacheFacade.txInfo().stats());
+        };
+        return args -> scheduler.scheduleWithFixedDelay(task, Duration.ofSeconds(60L));
+    }
+
+    @Bean
+    @Profile("!test")
+    CommandLineRunner mineBlocksOnRegtest(TaskScheduler scheduler, BitcoinClient bitcoinClient) {
+        if (!RegTestParams.get().equals(bitcoinClient.getNetParams())) {
+            return args -> {
+                log.debug("Not mining any blocks since client is not connected to regtest.");
+            };
+        } else {
+            Address eaterAddress = Address.fromString(RegTestParams.get(), "bcrt1q0lhr8js5rrhjl7hf6e80ns3pgk5tjswgg9um9t");
+
+            Runnable task = () -> {
+                try {
+                    List<Sha256Hash> blockHashes = bitcoinClient.generateToAddress(1, eaterAddress);
+                    log.info("Mined a new block on regtest: {}", blockHashes);
+                } catch (IOException e) {
+                    log.error("Error while mining regtest block: {}", e.getMessage());
+                }
+            };
+            return args -> scheduler.scheduleWithFixedDelay(task, Duration.ofSeconds(10L));
+        }
     }
 }