+++ changelog.md
@@ -10,8 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - module bitcoin-jsonrpc-client-autoconfigure: dependencies not included in anymore
 - module cln-grpc-client-autoconfigure: dependencies not included in anymore
 - module lnd-grpc-client-autoconfigure: dependencies not included in anymore
+- autoconfig property prefix of bitcoin jsonrpc cache changed from "jsonrpc.cache" to "jsonrpc-cache"
 
 ### Changes
+- improved default specs for bitcoin jsonrpc client caches
 - upgrade: update spring-boot from v3.1.0 to v3.1.1
 - upgrade: update cln-grpc-client-core from v23.5.1 to v23.5.2